+++ _skeleton.py
@@ -150,6 +150,7 @@ class ApodiktumSkeletonMod(loader.Module):
             "https://raw.githubusercontent.com/anon97945/hikka-mods/master/apodiktum_library.py",
             suspend_on_error=True,
         )
+        self.apo_lib.apodiktum_module()
 
     async def on_unload(self):
         return