+++ apo_migrator.py
@@ -133,6 +133,7 @@ class ApodiktumMigratorMod(loader.Module):
             "https://raw.githubusercontent.com/anon97945/hikka-mods/master/apodiktum_library.py",
             suspend_on_error=True,
         )
+        self.apo_lib.apodiktum_module()
 
     async def apomigratecmd(self, message: Message):
         """