+++ apo_migrator_class.py
@@ -137,6 +137,7 @@ class ApoAutoMigratorMod(loader.Module):
             "https://raw.githubusercontent.com/anon97945/hikka-mods/master/apodiktum_library.py",
             suspend_on_error=True,
         )
+        self.apo_lib.apodiktum_module()
         await self.apo_lib.migrator.auto_migrate_handler(self.__class__.__name__, self.strings("name"), self.changes, self.lookup("ApodiktumLib").config["auto_migrate"])
 
         # for test commands