+++ apo_python.py
@@ -89,6 +89,7 @@ class ApodiktumPythonMod(loader.Module):
             "https://raw.githubusercontent.com/anon97945/hikka-mods/master/apodiktum_library.py",
             suspend_on_error=True,
         )
+        self.apo_lib.apodiktum_module()
         self._phone = (await client.get_me()).phone
 
     @loader.owner