+++ apoinfo.py
@@ -274,6 +274,7 @@ class ApodiktumInfoMod(loader.Module):
             "https://raw.githubusercontent.com/anon97945/hikka-mods/master/apodiktum_library.py",
             suspend_on_error=True,
         )
+        self.apo_lib.apodiktum_module()
         self._me = await client.get_me()
 
     def _render_info(self) -> str: