+++ apolib_tester.py
@@ -64,6 +64,7 @@ class ApoLibTesterMod(loader.Module):
             "https://raw.githubusercontent.com/anon97945/hikka-mods/master/apodiktum_library.py",
             suspend_on_error=True,
         )
+        self.apo_lib.apodiktum_module()
 
     async def hellocmd(self, message: Message):
         await utils.answer(