+++ auto_update.py
@@ -200,6 +200,7 @@ class ApodiktumAutoUpdateMod(loader.Module):
             "https://raw.githubusercontent.com/anon97945/hikka-mods/master/apodiktum_library.py",
             suspend_on_error=True,
         )
+        self.apo_lib.apodiktum_module()
         asyncio.ensure_future(self._check_on_load(client))
 
     async def watcher(self, message: Message):