+++ autoreact.py
@@ -172,6 +172,7 @@ class ApodiktumAutoReactMod(loader.Module):
             "https://raw.githubusercontent.com/anon97945/hikka-mods/master/apodiktum_library.py",
             suspend_on_error=True,
         )
+        self.apo_lib.apodiktum_module()
 
     async def cautoreactcmd(self, message: Message):
         """