+++ dnd.py
@@ -279,6 +279,7 @@ class ApodiktumDNDMod(loader.Module):
             "https://raw.githubusercontent.com/anon97945/hikka-mods/master/apodiktum_library.py",
             suspend_on_error=True,
         )
+        self.apo_lib.apodiktum_module()
         self._ratelimit_afk = []
         self._ratelimit_pmbl = []
         self._ratelimit_pmbl_threshold = 10