+++ donators.py
@@ -171,6 +171,7 @@ class ApodiktumDonatorsMod(loader.Module):
             "https://raw.githubusercontent.com/anon97945/hikka-mods/master/apodiktum_library.py",
             suspend_on_error=True,
         )
+        self.apo_lib.apodiktum_module()
 
     async def cdonatorscmd(self, message: Message):
         """