+++ gtranslate.py
@@ -126,6 +126,7 @@ class ApodiktumGTranslateMod(loader.Module):
             "https://raw.githubusercontent.com/anon97945/hikka-mods/master/apodiktum_library.py",
             suspend_on_error=True,
         )
+        self.apo_lib.apodiktum_module()
         self._me = await client.get_me()
         self.tr = googletrans.Translator()