+++ msg_merger.py
@@ -209,6 +209,7 @@ class ApodiktumMsgMergerMod(loader.Module):
             "https://raw.githubusercontent.com/anon97945/hikka-mods/master/apodiktum_library.py",
             suspend_on_error=True,
         )
+        self.apo_lib.apodiktum_module()
 
     async def cmsgmergercmd(self, message: Message):
         """