+++ purge.py
@@ -200,6 +200,7 @@ class ApodiktumPurgeMod(loader.Module):
             "https://raw.githubusercontent.com/anon97945/hikka-mods/master/apodiktum_library.py",
             suspend_on_error=True,
         )
+        self.apo_lib.apodiktum_module()
 
     @staticmethod
     async def _purge_user_messages(chat, user_id, purge_count, message):