+++ pypng.py
@@ -107,6 +107,7 @@ class ApodiktumPyPNGMod(loader.Module):
             "https://raw.githubusercontent.com/anon97945/hikka-mods/master/apodiktum_library.py",
             suspend_on_error=True,
         )
+        self.apo_lib.apodiktum_module()
 
     async def pypngcmd(self, message: Message):
         """