+++ save_message.py
@@ -92,6 +92,7 @@ class ApodiktumSaveMessageMod(loader.Module):
             "https://raw.githubusercontent.com/anon97945/hikka-mods/master/apodiktum_library.py",
             suspend_on_error=True,
         )
+        self.apo_lib.apodiktum_module()
         self._id = (await client.get_me(True)).user_id
 
     async def smcmd(self, message: Message):