+++ show_viewer.py
@@ -99,6 +99,7 @@ class ApodiktumShowViewsMod(loader.Module):
             "https://raw.githubusercontent.com/anon97945/hikka-mods/master/apodiktum_library.py",
             suspend_on_error=True,
         )
+        self.apo_lib.apodiktum_module()
 
     async def svcmd(self, message: Message):
         """