+++ translations.py
@@ -119,6 +119,7 @@ class ApodiktumTranslationsMod(loader.Module):
             "https://raw.githubusercontent.com/anon97945/hikka-mods/master/apodiktum_library.py",
             suspend_on_error=True,
         )
+        self.apo_lib.apodiktum_module()
 
     async def ctranslationscmd(self, message: Message):
         """