+++ voicetools.py
@@ -468,6 +468,7 @@ class ApodiktumVoiceToolsMod(loader.Module):
             "https://raw.githubusercontent.com/anon97945/hikka-mods/master/apodiktum_library.py",
             suspend_on_error=True,
         )
+        self.apo_lib.apodiktum_module()
         self._id = (await client.get_me(True)).user_id
 
     async def cvoicetoolscmd(self, message: Message):